+++ include/web_pages.h
@@ -3,7 +3,7 @@
 
 #include <Arduino.h>
 
-const char *const VERSION = "1.10.2";
+const char *const VERSION = "1.10.3";
 
 // --- HTML Content ---